requirements.txt
numpy==1.20.1
pandas==1.2.3
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
